Moved license and recompiles on Mac
I had the GPL v2 in the root of the directory structure. I moved it to
the headers of the file to be more explicit and removed it from the
root of the directory structure. I also reran the setup_FASPL script to
make sure that I didn't break anything.